# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  model.SureSurePackage:
    properties:
      amount:
        type: number
      created_date:
        type: string
      duration:
        type: integer
      id:
        type: integer
      is_active:
        type: integer
      ordered:
        type: integer
      package_name:
        type: string
      package_price:
        type: number
      quota_limit:
        type: integer
      updated_date:
        type: string
    type: object
  model.SureSureRoom:
    properties:
      created_date:
        type: string
      id:
        type: integer
      line_group_id:
        type: string
      list_bank:
        type: string
      min_receive:
        type: number
      qr_token:
        type: string
      quota_used:
        type: integer
      room_name:
        type: string
      show_recipient:
        type: boolean
      show_transferor:
        type: boolean
      updated_date:
        type: string
      user_id:
        type: integer
    type: object
  model.SureSureTransaction:
    properties:
      amount:
        type: number
      created_date:
        type: string
      cstid:
        type: string
      id:
        type: integer
      line_group_id:
        type: string
      line_user_id:
        type: string
      message:
        type: string
      proxy_account_no:
        type: string
      qr_code:
        type: string
      receive_account_no:
        type: string
      receive_bank_code:
        type: string
      receive_name:
        type: string
      receive_name2:
        type: string
      ref_no:
        type: string
      ref1:
        type: string
      ref2:
        type: string
      rquid:
        type: string
      sender_account_no:
        type: string
      sender_bank_code:
        type: string
      sender_name:
        type: string
      sender_name2:
        type: string
      status:
        type: string
      status_code:
        type: string
      trans_date:
        type: string
      trans_time:
        type: string
      txid:
        type: string
      updated_date:
        type: string
      user_id:
        type: integer
    type: object
...
  util.APIResponse:
    properties:
      data: {}
      message:
        type: string
      status_code:
        type: integer
      timestamp:
        type: string
    type: object
info:
  contact: {}
  description: Public API for SureSure services
  title: SureSure Public API
  version: "1.0"
paths:
...
  /package/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: Package payload
        in: body
        name: package
        required: true
        schema:
          $ref: '#/definitions/model.SureSurePackage'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.APIResponse'
      summary: Create package
      tags:
      - Package
  /package/delete/{id}:
    delete:
      parameters:
      - description: Package ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.APIResponse'
      summary: Delete package by ID
      tags:
      - Package
  /package/get:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.APIResponse'
      summary: List packages
      tags:
      - Package
  /package/get/{id}:
    get:
      parameters:
      - description: Package ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.APIResponse'
      summary: Get package by ID
      tags:
      - Package
  /package/update:
    put:
      consumes:
      - application/json
      parameters:
      - description: Package payload
        in: body
        name: package
        required: true
        schema:
          $ref: '#/definitions/model.SureSurePackage'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.APIResponse'
      summary: Update package
      tags:
      - Package
  /room2/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: Room payload
        in: body
        name: room
        required: true
        schema:
          $ref: '#/definitions/model.SureSureRoom'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.APIResponse'
      summary: Create room
      tags:
      - Room
  /room2/delete/{id}:
    delete:
      parameters:
      - description: Room ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.APIResponse'
      summary: Delete room by ID
      tags:
      - Room
  /room2/get:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.APIResponse'
      summary: List rooms
      tags:
      - Room
  /room2/get/{id}:
    get:
      parameters:
      - description: Room ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.APIResponse'
      summary: Get room by ID
      tags:
      - Room
  /room2/howto/{id}/{user_id}:
    get:
      parameters:
      - description: Room ID
        in: path
        name: id
        required: true
        type: integer
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.APIResponse'
      summary: HowTo action for room
      tags:
      - Room
  /room2/update:
    put:
      consumes:
      - application/json
      parameters:
      - description: Room payload
        in: body
        name: room
        required: true
        schema:
          $ref: '#/definitions/model.SureSureRoom'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.APIResponse'
      summary: Update room
      tags:
      - Room
  /transaction/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: Transaction payload
        in: body
        name: transaction
        required: true
        schema:
          $ref: '#/definitions/model.SureSureTransaction'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.APIResponse'
      summary: Create transaction
      tags:
      - Transaction
  /transaction/delete/{id}:
    delete:
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.APIResponse'
      summary: Delete transaction by ID
      tags:
      - Transaction
  /transaction/get:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.APIResponse'
      summary: List transactions
      tags:
      - Transaction
  /transaction/get/{id}:
    get:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.APIResponse'
      summary: Get transaction by user ID
      tags:
      - Transaction
  /transaction/update:
    put:
      consumes:
      - application/json
      parameters:
      - description: Transaction payload
        in: body
        name: transaction
        required: true
        schema:
          $ref: '#/definitions/model.SureSureTransaction'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.APIResponse'
      summary: Update transaction
      tags:
      - Transaction
...
schemes:
- http
- https
swagger: "2.0"